#106
Weer toegevoegd bij predicted data en in de predictiondata model een double gemaakt van de temperatuur

+++ TrashDashboard/TrashDashboard/Models/PredictedTrash.cs
@@ -1,9 +1,9 @@
-﻿namespace TrashDashboard.Models
+﻿    namespace TrashDashboard.Models
 {
     public class PredictedTrash
     {
         public DateTime Date { get; set; }
-        public int PredictedTemprature { get; set; }
+        public double PredictedTemprature { get; set; }
         public bool IsHoliday { get; set; }
         public int PredictedWasteCount { get; set; }
     }
